Make trans algorithm look a lot closer to Monocurl 1.0

+++ crates/stdlib/src/scene.rs
@@ -98,9 +98,9 @@ pub async fn initial_background(
         [(
             "color",
             value_list([
-                Value::Integer(0),
-                Value::Integer(0),
-                Value::Integer(0),
+                Value::Integer(1),
+                Value::Integer(1),
+                Value::Integer(1),
                 Value::Integer(1),
             ]),
         )],
